refactor(backend): refactor package structure
