+++ source/did-ca-server/src/main/java/org/omnione/did/cas/v1/admin/controller/AdminTestController.java
@@ -22,8 +22,8 @@ import org.omnione.did.base.db.domain.CertificateVc;
 import org.omnione.did.base.util.BaseMultibaseUtil;
 import org.omnione.did.cas.v1.agent.service.CertificateVcQueryService;
 import org.omnione.did.cas.v1.agent.service.StorageService;
-import org.omnione.did.cas.v1.common.dto.admin.test.EmptyResDto;
-import org.omnione.did.cas.v1.common.dto.admin.test.SendCertificateVcReqDto;
+import org.omnione.did.cas.v1.common.dto.EmptyResDto;
+import org.omnione.did.cas.v1.admin.dto.cas.SendCertificateVcReqDto;
 import org.omnione.did.data.model.did.DidDocument;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
