+++ source/did-ca-server/src/main/java/org/omnione/did/cas/v1/admin/service/AdminManagementService.java
@@ -27,8 +27,8 @@ import org.omnione.did.cas.v1.admin.dto.admin.RegisterAdminReqDto;
 import org.omnione.did.cas.v1.admin.dto.admin.ResetPasswordByRootReqDto;
 import org.omnione.did.cas.v1.admin.dto.admin.ResetPasswordReqDto;
 import org.omnione.did.cas.v1.admin.dto.admin.VerifyAdminIdUniqueResDto;
-import org.omnione.did.cas.v1.common.service.query.AdminQueryService;
-import org.omnione.did.cas.v1.common.dto.admin.test.EmptyResDto;
+import org.omnione.did.cas.v1.admin.service.query.AdminQueryService;
+import org.omnione.did.cas.v1.common.dto.EmptyResDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
