+++ source/did-ca-server/src/main/java/org/omnione/did/cas/v1/admin/service/CasManagementService.java
@@ -36,7 +36,6 @@ public class CasManagementService {
     private final CasQueryService casQueryService;
     private final StorageService storageService;
 
-    @RequestMapping(value = "/ca/info", method = RequestMethod.GET)
     public GetCasInfoReqDto getCasInfo() {
         Cas existingCas = casQueryService.findCas();
