+++ source/did-ca-server/src/main/java/org/omnione/did/cas/v1/admin/service/query/AdminQueryService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.omnione.did.cas.v1.common.service.query;
+package org.omnione.did.cas.v1.admin.service.query;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
